feat: create personal account - finalize pin code screen
